[FE] fix : 네브바, 글씨체, alert문구, 라우터가드, 버튼 CSS
Former-commit-id: ed2237f2d2b8346c284e91133437eabb811480ef

+++ frontend/src/store/modules/accounts.js
@@ -1,6 +1,5 @@
 import axios from 'axios'
 import drf from '@/api/api'
-import router from '@/router/index'
 
 export default {
 	state: {
@@ -15,6 +14,9 @@ export default {
 		mainImage: [],
 		selectList: [],
 		convertList: [],
+		preventInfoPage: false,
+		preventSignupPage: false,
+		preventSelectPage: false,
 	},
 	getters: {
 		authHeader: state => ({ Authorization: `Bearer ${state.token}` }),
@@ -29,6 +31,9 @@ export default {
 		selectList: state => state.selectList,
 		isExisted: state => state.isExisted,
 		convertList: state => state.convertList,
+		preventInfoPage: state => state.preventInfoPage,
+		preventSelectPage: state => state.preventSelectPage,
+		preventSignupPage: state => state.preventSignupPage,
 	},
 	mutations: {
 		SET_TOKEN(state, token) {
@@ -67,6 +72,15 @@ export default {
 		SET_CONVERT_LIST(state, data) {
 			state.convertList = data
 		},
+		PREVENT_INFO_FORM_PAGE(state, value) {
+			state.preventInfoPage = value
+		},
+		PREVENT_SELECT_PAGE(state, value) {
+			state.preventSelectPage = value
+		},
+		PREVENT_SIGNUP_PAGE(state, value) {
+			state.preventSignupPage = value
+		},
 	},
 	actions: {
 		saveToken({ commit }, token) {
@@ -84,14 +98,14 @@ export default {
 				},
 			})
 				.then(() => {
-					router.push({ name: 'SelectImage' })
+					window.location.href = 'http://localhost:3002/select'
 				})
 				.catch(() => {
 					alert('서비스가 비정상적입니다. 다시 시도해주세요.')
 				})
 		},
 
-		doubleCheck({ getters, dispatch }, userNickName) {
+		doubleCheck({ getters, dispatch, commit }, userNickName) {
 			axios({
 				headers: getters.authHeader,
 				url: drf.auth.nickNameCheck(getters.userId),
@@ -102,6 +116,7 @@ export default {
 			})
 				.then(() => {
 					dispatch('nickNameForm', userNickName)
+					commit('PREVENT_SIGNUP_PAGE', true)
 				})
 				.catch(() => {
 					alert('사용할 수 없는 닉네임입니다.')
@@ -140,7 +155,7 @@ export default {
 				},
 			})
 				.then(() => {
-					router.push({ name: 'Main' })
+					window.location.href = 'http://localhost:3002/main'
 					dispatch('getMainBasedOnSelected')
 				})
 				.catch(() => {
@@ -175,11 +190,13 @@ export default {
 				method: 'get',
 			})
 				.then(() => {
-					dispatch('removeToken')
-					commit('REMOVE_TOKEN')
-					localStorage.setItem('vuex', '')
-					confirm('로그아웃하시겠습니까?')
-					router.push({ name: 'Login' })
+					const next = confirm('로그아웃하시겠습니까?')
+					if (next) {
+						dispatch('removeToken')
+						commit('REMOVE_TOKEN')
+						window.location.href = 'http://localhost:3002/'
+						localStorage.setItem('vuex', '')
+					}
 				})
 				.catch(() => alert('서비스가 비정상적입니다. 다시 시도해주세요.'))
 		},
